Fix PHPstan warnings

+++ src/DaPigGuy/PiggyCustomEnchants/enchants/miscellaneous/ToggleableEffectEnchant.php
@@ -53,8 +53,9 @@ class ToggleableEffectEnchant extends ToggleableEnchantment
             }
         }
         $player->getEffects()->remove($this->effect);
+
         $amplifier = $this->extraData["baseAmplifier"] + $this->extraData["amplifierMultiplier"] * $level;
-        $player->getEffects()->add(new EffectInstance($this->effect, 2147483647, min($amplifier, 255), false));
+        $player->getEffects()->add(new EffectInstance($this->effect, 2147483647, (int)(min($amplifier, 255)), false));
     }
 
     public function getUsageType(): int
